Added my Razor stuffs
The old RazorViewEngine project could probably be removed.. the new Html
is for handling Html formatting in the WebApi, it supports different
types of parsers (as long as they are implemented of course). The .Razor
project is for parsing Razor to Html.. I also added a ASP.Net Web Api
project for showing how to use Razor stuffs.. it should of course be
removed later.. in the final release.