Fixed a potential bug regarding ViewLocator and ViewParser
The problem with the code was: If one of the ViewParser or ViewLocator
was null, it replace both from the Dependency Resolver, but it only
passes them to the HtmlMediaTypeViewFormatter if both was resolved.

The fix will now take any of them..

To decide:  We can add a check that make sure both must exists in the
IoC container and throw an exception if they don't, or let them mix
between using the global config's DefaultXXXX and also register one of
them in a IoC containter.. I think if one is only registred in the IoC
but the other is specified as a default, the code should still work.