Added HttpError page, added a start page
I have added a HttpError, so when a controller throws an error, the
HttpError.cshtml page will be used. Works well in Crome, but IE will
show its own 500 Error page. I also added a start page, so when starting
up the MvcWebApiSiteTest, the Index.cshtml page will be displayed. The
idea of that start page is to give a few intro about how to use this
Razor feature.